move restmail out of utilities for codeclimate
